2016-04-08 Finished fixing issue #32, started issue #6

+++ app/src/main/java/net/innit/drugbug/fragment/AddDoseFragment.java
@@ -45,7 +45,6 @@ import java.io.File;
 import java.text.DateFormat;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
-import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
 import java.util.Locale;
@@ -123,7 +122,7 @@ public class AddDoseFragment extends DialogFragment {
     @Nullable
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.activity_add_edit, container, false);
+        View view = inflater.inflate(R.layout.fragment_add_edit, container, false);
 
         setupViews(view);
         if (savedInstanceState != null) {
@@ -185,7 +184,9 @@ public class AddDoseFragment extends DialogFragment {
     public void onCancel(DialogInterface dialog) {
         super.onCancel(dialog);
         this.dismiss();
-        tempPath.delete();
+        if (tempPath != null && tempPath.exists()) {
+            tempPath.delete();
+        }
     }
 
     private void onCreateEdit(Bundle bundle) {
@@ -327,6 +328,8 @@ public class AddDoseFragment extends DialogFragment {
                 db.updateMedication(medication);
             }
 
+            doseItem.setMedication(medication);
+
             doseItem.setDosage(mDosage.getText().toString());
 
             doseItem.setReminder(mReminder.isChecked());
@@ -345,21 +348,11 @@ public class AddDoseFragment extends DialogFragment {
                             nextDose = updateFutureDose(nextDose, futureDose);
                         }
                     } else {
-                        // TODO: 4/2/16 wrap this up into db.createNextFuture
-                        DoseItem futureItem = doseItem;
-                        int numFutureDoses = Integer.parseInt(Settings.getInstance().getString(Settings.Key.NUM_DOSES));
-                        Calendar calendar = Calendar.getInstance();
-                        for (int i = 0; i < numFutureDoses; i++) {
-                            calendar.setTime(futureItem.getDate());
-                            if (i > 0) {
-                                // Add interval to last date
-                                calendar.add(Calendar.SECOND, (int) db.getInterval(medication.getFrequency()));
-                            }
-                            futureItem = new DoseItem(medication, calendar.getTime(), doseItem.isReminderSet(), false, doseItem.getDosage());
-                            futureItem = db.createDose(futureItem);
+                        db.createDose(doseItem).setOrKillAlarm(getActivity());
 
-//                            handleReminder(futureItem);
-                            futureItem.setOrKillAlarm(getActivity());
+                        int numFutureDoses = Integer.parseInt(Settings.getInstance().getString(Settings.Key.NUM_DOSES));
+                        for (int i = 0; i < numFutureDoses-1; i++) {
+                            doseItem.getMedication().createNextFuture(getActivity()).setOrKillAlarm(getActivity());
                         }
                     }
 
@@ -372,7 +365,6 @@ public class AddDoseFragment extends DialogFragment {
                         } else {
                             intent.putExtra(FILTER_MED, medFilter);
                         }
-                        // TODO: 4/2/16 May need some jiggery pokery here to account for changingto FILTER_ALL if sort order is by date
                         intent.putExtra(SORT_MED, medSortOrder);
                         startActivity(intent);
                         dismiss();
@@ -417,7 +409,7 @@ public class AddDoseFragment extends DialogFragment {
         if (thisDose.getId() != doseItem.getId()) {
             // Only do the setting if this is not the dose currently being edited
             if (thisDose.getDate().getTime() > origDate.getTime()) {
-                thisDose.setDate(nextDose.nextDate(getActivity()));
+                thisDose.setDate(nextDose.getMedication().getNextDate(getActivity(),nextDose).getTime());
                 nextDose = thisDose;
             }
             thisDose.setDosage(doseItem.getDosage());
